pyproject.toml
[project]
name = "ppt-to-txt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "fastapi>=0.115.12",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pydantic-ai-slim[openai]>=0.2.16",
    "pytest>=8.4.0",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
